Replace named groups more reliably.
If a replaced param's value contained ":" (a colon, the named group
prefix) then the regex would find it next time and replace the wrong bit
of the URL.

I considered prefixing the regex with a non-remembering match for the
"/" that should always be in front of the ":" ... but it would still
fail when a replacement value *started* with a ":".

In the end, I decided to drop the regex's and just iterate the path
segments looking for named groups. Much easier.

I think this should handle any group name now too (the AngularJS docs
suggests a name is not restricted in any way).